# Conflicts:
#	ODIN-SampleProject/Assets/ODIN-Sample/Scripts/Runtime/ODIN/OdinAudioFilterSettingsController.cs

+++ ODIN-SampleProject/Assets/4Players/ODIN/Runtime/Events/UnityCreatedMediaObject.cs
@@ -9,7 +9,7 @@ using OdinNative.Odin.Room;
 namespace OdinNative.Unity.Events
 {
     [Serializable]
-    public class UnityCreatedMediaObject : UnityEvent<string, ulong, ushort>
+    public class UnityCreatedMediaObject : UnityEvent<string, ulong, long>
     {
     }
 }
